[CBRD-25068] Increase session timeout value for isolation test due to performance degradation from memory monitoring (#673)
Due to memory monitoring, the performance of the debug mode has decreased, which is expected. In the Isolation debug test, the timeout for a session is set to 30 seconds, causing some queries to exceed this limit and resulting in failed test cases. To address this issue, the 'SHORT_DURATION_DEFAULT' value is increased from 30 seconds to 100 seconds.