[CBRD-25352] Add a user schema to your stored procedures for consistency with other objects.
http://jira.cubrid.org/browse/CBRD-25352

When calling built-in PL/SQL from CTE, you need to prefix with 'dbms_output.' to ensure normal execution. 
For example, 'call get_line;' -> 'call dbms_output.get_line;'

When the PR is merged, it affects the develop branch into which the existing phase-0 was merged.
Therefore Until phase-1 is merged, we have modified the CTE to temporarily remove the "DBMS_OUTPUT" from the Built-in PL/CSQL in the catch block and run it once more. 

We plan to remove it again when the phase-1 feature branch is merged in the future.